docs(ast): add more docs for JS expressions, declarations, and module AST types (#8800)
Tried to add documentation and examples for as many of the remaining AST methods as I could. There are still around 30 or so remaining in this file, but this PR adds around 50 documentation comments.

I tried my best to check that each comment is correct, but I was inferring from the source and usage, so it may not be 100% accurate. Please take a quick look over the comments and see if it looks right.