fix(linter): unicorn/switch-case-braces mangles code when applying fix (#8758)
Addresses #8491.

Essentially, the fix was ensuring that we track the indentation of a
`StatementExpression` or `BreakStatement` when we apply a fix for this
lint rule, and to make sure we add these statements on a new line with
the proper indentation. I also made sure we're indenting the closing
case brackets correctly.

To do this, I added a `get_preceding_indent_str` fn to the `ast_utils`
that handles getting the preceding indentation of a `Span` in a
`source_text` &str. It returns an Option in case no indentation is
found, or if the statement is not the only one on a given line.

This new fn was useful when addressing this particular bug, but I figure
it might be useful for other fixer use cases, too. I added some
documentation for this fn for clarity.