refactor(syntax): add `#[ast]` attr to semantic ID types (#8867)
Add `#[ast]` attr to `ScopeId`, `SymbolId` and `ReferenceId` types. This makes them an "official" part of the AST, and removes the hacky "special case" logic from `oxc_ast_tools` which was previously working around that they weren't registered as part of the AST.

This uses the `#[ast(foreign = ...)]` feature introduced in #8866 to inform `oxc_ast_tools` about the foreign `NonMaxU32` type, enabling the ID types to be parsed and understood.