feat(ast_tools): generate mapping of trait name to crate in `oxc_ast_macros` (#8870)
`oxc_ast_macros` contains a function which takes the name of a trait, and returns the crate that trait is defined in.

Instead of having that info hard-coded in 2 places, codify it in generator for that trait in `oxc_ast_tools`, and codegen this function in `oxc_ast_macros` instead.