refactor(ast): generate `Visit` methods in order of type definitions (#8891)
Pure refactor. Re-order the methods in generated `Visit` and `VisitMut` traits to be in same order they appear in source code.

This simplifies and improves performance of the codegen considerably. Previously it walked the AST types recursively, with loads of checks "did I generate visitor for this type already?". Now it's just a single loop over the type definitions.

I've double-checked that this does not result in new/missing visit methods. This change purely results in the same methods being output in a different order.