refactor(ast): estree `via` on struct fields implement `From` (#8898)
Align the behavior of custom serialization to ESTree using the `#[estree(via)]` attribute. Change its behavior on struct fields to match structs - the "via" type must implement `From` to convert from the original type to itself.

The future goal is to use these same "via" types to go back in the other direction (JS -> Rust), so we need to have a casting system.