fix(minifier): should not merge conditional function calls if referencing the function has a side-effect (#8922)
`a ? b(c, d) : b(e, d)` should not be compressed to `b(a ? c : e, d)` if `b` has a side effect.
Because `b` may update `a` variable.

For example, compressing the following code changes the behavior.
```js
var foo = () => console.log('foo')
var bar = () => { foo = () => console.log('foo2') }
bar() ? foo(1, 2) : foo(3, 2) // outputs 'foo'

// was compressed into
var foo = () => console.log('foo')
var bar = () => { foo = () => console.log('foo2') }
foo(bar() ? 1 : 3, 2); // outputs 'foo2'
```

We can improve this if we know that `foo` is not updated after it was declared (i.e. `foo` is a const). But I didn't implement it for now.

<details>
<summary>evaluation steps in detail</summary>

Before the compression, the code is evaluated in the following steps:
1. Evaluate `a`
1. Evaluate `b`
1. Evaluate `c` / `e`
1. Evaluate `d`
1. Evaluate `b(c,d)` / `b(e,d)`

After the compression, the code is evaluated in the following steps:
1. Evaluate `b`
1. Evaluate `a`
1. Evaluate `c` / `e`
1. Evaluate `d`
1. Evaluate `b(c,d)` / `b(e,d)`

The steps of 1 and 2 are reversed.

</details>

**References**
- [Spec of conditional expression](https://tc39.es/ecma262/multipage/ecmascript-language-expressions.html#sec-conditional-operator-runtime-semantics-evaluation)
- [Spec of call expression](https://tc39.es/ecma262/multipage/ecmascript-language-expressions.html#sec-function-calls-runtime-semantics-evaluation)