chore(build): bump MSRV to 1.81.0 (#8931)
As noted in #8929, many Oxc crates won't compile at present with our MSRV of 1.80.0, due to the liberal use of `#[expect]` attribute, which wasn't stabilized until 1.81.0.

It becomes a lint error in 1.81.0 to use `#[allow]`, so it'd seem to me pretty pointless to convert all the `#[expect]` attrs to `#[allow]` to honour our MSRV now, when we'll have to change them back to `#[expect]` again shortly.

So I propose that we bump the MSRV to 1.81.0 early instead. Next minor version of Rust is out in 2 weeks anyway.