perf(ast): assume `serde_json` output is valid UTF8 string (#8928)
`serde_json` produces valid UTF8 output, so when converting serialized JSON to string, there's no need to check the output for UTF8 validity. Skipping that check is a speed-up.

Additionally, add a method to test the serializer without producing any output, and use it in conformance. This will hopefully speed up conformance a little.