refactor(ast)!: remove `PrivateInExpression::operator` field (#9041)
Remove `PrivateInExpression::operator` field.

This field had no purpose beyond getting correct serialization to ESTree - it was always `BinaryOperator::In`, and no other value for `operator` makes sense in this position. So remove the field from the Rust type, and produce the desired output in ESTree AST using `#[estree(add_fields)]` instead.

This follows on after changes made to serialization of `PrivateInExpression` in #9033.