test(ast/estree): bump `acorn-test262` submodule (#9047)
Bump `acorn-test262` submodule to integrate commit https://github.com/oxc-project/acorn-test262/commit/dcb070006436b9f7783626c0741d5693953035c6 which fixes Acorn's output for 1 of the ESTree conformance tests which was previously failing.

The problem on that test case wasn't Oxc, the problem was that we were running Acorn with the wrong options.