feat(ast_tools): introduce meta types (#9117)
All AST types are tagged `#[ast]` to communicate to `oxc_ast_tools` that they're part of the AST.

Introduce `#[ast_meta]` attribute to define "meta types". Meta types are not part of the AST, but are used by `oxc_ast_tools` in generating code for AST types. Information about meta types can be communicated to `oxc_ast_tools` via further attributes on the types, same as for types marked `#[ast]`.