refactor(ast/estree): simplify serializing `FormalParameters` (#9195)
Simplify the custom `Serialize` impl for `FormalParameters`.

This removes the last `#[derive(Serialize)]` in `oxc_ast` crate, which means we can turn off `serde`'s `derive` feature for this crate.

Unfortunately, `oxc_ast` depends on `oxc_span`, and that crate still uses `#[derive(Deserialize)]`. Therefore `oxc_ast` still has a transitive dependency on `serde-derive`. But this is a step to removing it, which when we do will greatly improve compile times.