refactor(span): implement `Serialize` manually for `Atom` (#9196)
The `Serialize` impl for `Atom` is trivial, so there's no need to use `#[derive(Serialize)]` on it - just implement it by hand.

This is another step towards removing `serde-derive` from `oxc_span` and `oxc_ast`, which when complete will greatly improve compile times.

Also shorten the code for `impl Serialize for CompactStr`. That change is pure style refactor.