feat(ast_tools): add `#[builder(default)]` attribute for structs and enums (#9203)
Add a `#[builder(default)]` attribute which can be used on structs or enums to indicate that AST builder methods should not require this field.

Use that attribute on `ScopeId`, `SymbolId` and `ReferenceId`, to replace the hard-coded list of semantic ID types in codegen for `AstBuilder`.

Does not affect generated code, only the mechanisms by which it's generated.