feat(ast_tools): add `#[builder(skip)]` attribute for structs and enums (#9205)
Add a `#[builder(skip)]` attribute which can be used on structs or enums to indicate no `AstBuilder` methods should be generated for this type.

Use that attribute on `Span`, to replace the hard-coded black list in codegen for `AstBuilder`.

Does not affect generated code, only the mechanism by which it's generated.