feat(linter): support `import-x` plugin name (#9074)
closes #8779

The `import-x` rules are the reimplementation of the `import` rules. No rules added or removed.
Some eslint-plugins/configs are handling the plugin the same way.

Here a snippet from a config which `next-auth` is using:
https://github.com/antfu/eslint-config/blob/998887cedb0b09c5b95d08602689a2c86e6662c7/src/factory.ts#L56-L67

Could not find more places where some mapping is happening. Please have a look because I did not test it.