feat(transformer/statement-injector): add an assertion to check if it still contains statements that don't inject (#9063)
When there are still statements that aren't inserted after the visitor ends, it means statements inserted with an incorrect address or outdated address (e.g., the original statement has mutated to a new statement).

This PR adds an assertion to expose this potential bug as soon as possible on the development side.