feat(oxlint): add `--experimental-nested-config` option (#9152)
- part of https://github.com/oxc-project/oxc/issues/7408

Since the nested configuration work is too big to fit in one PR or even a stack of PRs, I suggest that we make it temporarily an opt-in CLI flag. Once we feel confident in the implementation replacing the existing code, we can remove the flag and make it true by default.