fix(wasm): transfer AST to JS as JSON string (#9259)
In [@oxc-parser/wasm](https://www.npmjs.com/package/@oxc-parser/wasm) NPM package, transfer the AST to JS as a JSON string, and parse it to a `Program` object on JS side.

This fixes a couple of bugs - previously `RegExp`s and `BigInt`s were deserialized as `null`.

The mechanism for doing this is pretty hacky. After `wasm-pack` generates the JS binding files, a script modifies these files to add a getter for `program` property on `ParseResult`. That getter deserializes the JSON to a `Program`.

The hackyness is not ideal, but I'm not familiar enough with WASM build systems to see if there's a better way. This does at least seem to work!