refactor(transformer/decorators): get statement address with `GetAddress::address` (#9264)
Follow-on after #9062. Pure refactor.

In all these cases, `class` / `export` is a `Box`, so we can use `class.address()` to get the address, instead of `Address::from_ptr(class.as_ref())`.

These 2 both do exactly the same thing, but personally I think it's preferable to avoid `Address::from_ptr` where we can. `Address::GetAddress` always produces a valid, permanent `Address`, whereas `Address::from_ptr` is a bit of a hack and can give you an unreliable `Address` in some cases. If we only use `Address::from_ptr` where we *have* to, it's easier to see where we have a potential sources of bugs.