refactor(transformer): do not use `&mut` when `&` ref will do (#9266)
#9253 enabled clippy's `needless_pass_by_ref_mut` rule, but kept it disabled for the `oxc_transformer` crate.

This rule is really good as using `&` instead of `&mut` may allow the compiler to make more optimizations in some cases, as `&` guarantees that value is not changed during the function call which takes `&`.

Enable this lint and fix all the warnings it highlights.

The diff is large, but it's purely replacing `&mut` with `&` in lots of places.