fix(wasm): transfer AST to JS as JSON string in `oxc-wasm` (#9269)
Same as #9259, but for `oxc-wasm` package.

This fixes a couple of bugs - previously `RegExp`s and `BigInt`s were
deserialized as `null`.

Only differences from the implementation in
[@oxc-parser/wasm](https://www.npmjs.com/package/@oxc-parser/wasm) are:

1. Getter for `ast` does not cache the result of `JSON.parse`. I
couldn't quite get my head around Playground code, but I *think* the
`Oxc` class instance is being used as a singleton, and therefore the
value may change as the user updates the source code.

2. JSON is pretty-printed. This is so that when "Raw" is ticked, the
JSON can be displayed exactly as it came out of the parser (rather than
stringifying the AST object again).