test(ast): ESTree conformance tester skip cases including infinite numbers (#9280)
Skip test cases where source code contains a number too large to fit in
an `f64`, which is treated as `Infinity`.

The JSON that `ESTree` serializer produces is fine - `Infinity` is
serialized as `1e+400`, which `JSON.parse` interprets correctly as
`Infinity`. But `serde`, as used in ESTree conformance tester, can't
deal with this which makes these tests erroneously fail. So skip them.