test(ast/estree): ESTree conformance tester ignore fixtures with hashbangs (#9283)
ESTree conformance tester ignore 6 fixtures where there's a hashbang.

The `Program::hashbang` field is not part of ESTree (ESTree treats it as a comment). But in my opinion, this field is useful, and we should diverge from ESTree to include it on our JS-side AST. The addition of this field won't break anyone's code.

Bump `acorn-test262` to include commit https://github.com/oxc-project/acorn-test262/commit/3ab34526674f12701b98022fa1987a084ad3f356 which adds a `hashbang: null` field to `Program` in the Acorn JSON files.

This PR prepares for checking our AST aginst Acorn's with a direct JSON-to-JSON comparison, ensuring the addition of this extra field doesn't cause all the tests to fail.