feat(ast/estree): enable serialization without TS fields (#9285)
Add the ability to serialize to a completely ESTree-compatible AST, omitting all TypeScript-related fields.

* `Program::to_estree_ts_json` -> AST with TypeScript fields.
* `Program::to_estree_js_json` -> AST without TypeScript fields.

The AST generated by `to_estree_js_json` completely aligns with the output of `rollup/parseAst`.

The ESTree conformance tester uses the new `Program::to_pretty_estree_js_json` method, and tests our ESTree AST against Acorn's with a direct comparison of JSON strings - rather than the previous complicated process of parsing the JSON, amending the AST, stringifying to JSON again, and then comparing.

This is both more reliable, and also much faster.

100% of tests passing.

We'll need to decide what to do with this. We could either:

1. Add an option to `parseSync` to get a JS-only AST. or
2. Make returning JS-only AST the default, if the source type is JS/JSX.

This PR does neither - it makes no changes to the NAPI or WASM functions. So no breaking changes. I think we can merge it, and decide how to use this feature later.