refactor(data_structures): reduce scope of `unsafe` blocks (#9316)
The new lint in Rust 1.85.0 `unsafe_op_in_unsafe_fn` is good - even in an unsafe function, it's good to make it clear what the unsafe operations are.

But the auto-fix for this lint was not good - it just encased entire function body in an `unsafe {}` block. If anything this has the opposite effect - making it *less* clear where unsafe invariants are generated and satisfied.

In `oxc_data_structures`, reduce the scope of these `unsafe {}` blocks, so they cover only the unsafe operations.