fix(data_structures): stack types correctly report allocation size if allocation failure during grow (#9317)
If an allocation failure during growth in `Stack`, `NonEmptyStack` and `SparseStack`, they would report the failure with details of the old allocation size, rather than the new one. Fix that.

This bug only affected error reporting, not soundness - and allocation failures should never happen anyway. But still, better to report it correctly.