perf(ast/estree): speed up building UTF8-UTF16 translation table with SIMD (#9359)
Use SIMD operations to speed up building the table of UTF8-UTF16 offset translations.

Previously, we looped through the source byte-by-byte, checking if each byte is ASCII, one-by-one. Instead, use SIMD to check blocks of 32 bytes at a time, and only drop down to byte-by-byte processing if the quick check does find any non-ASCII bytes in that block.

Building the table is about 7x faster in our benchmark after this PR.

Note: There are no explicit SIMD operations involved in the code. Just the code is written so that the compiler can auto-vectorize it.

We could get even better performance if we did use SIMD proper - on x86_64 with AVX-512 or Apple M4, could process blocks of 64 bytes in one go, and with less instructions. But that involves a lot of complications, so I think this is a decent compromise for now.