refactor(ast/estree): implement custom serializers as meta types (#9373)
Refactor. For structs and enums with custom serializers, use `#[estree(via = Converter)]` instead of `#[estree(custom_serialize)]`. Converter must be a meta type defined with `#[ast_meta]`.

This reduces the number of `#[estree]` attributes on the AST Rust type definitions, and makes it consistent with struct field converters. All custom conversion is now performed via meta types.

A side effect is that it removes TS type def for `FormalParameters`, which is not part of ESTree AST, so shouldn't be included in TS typings.