ci(benchmarks): exclude parser from isolated declarations benchmark timing (#9388)
Similar to other benchmarks, in benchmark for isolated declarations, only time the component under test - the actual generation of isolated declarations file. Previously, the benchmark also included parsing.

This removes a source of variance from the benchmark and "zooms in" on performance of the thing itself. It'll be easier to see improvements/degradations in performance.