perf(ast/estree): reduce overhead serializing static strings (#9396)
By default, strings have to be processed to ensure they contain no characters which need to be escaped in JSON, and escape those characters when they do. This is quite slow. But many strings the codegen produces are statically known not to require escaping e.g. `"IdentifierReference"` in `{"type":"IdentifierReference",...}`.

Introduce `JsonSafeString` type. Wrapping a string in this type asserts that the string definitely doesn't need to be escaped, and serializer can skip checks.

In codegen for `ESTree` impls, check if static strings need escaping, and wrap them in `JsonSafeString` if they don't.