fix(linter): support more cases for no_redundant_constructor_init (#9364)
Follow-up from https://github.com/oxc-project/oxc/pull/9299

As pointed out in discussions there, the initial implementation doesn't
handle assignments nested in blocks as only the body's statements were
being checked.

This instead checks for any references to each public param, checks it
is a read (which it would be in the right-side of an assignment) and
then if it's part of an assignment to `this`.

Adds tests for these new cases, previous tests still pass.