fix(ast_tools): fix calculation of niches for `Option`s (#9422)
Fix a bug in memory layout calculations in `oxc_ast_tools`. It was determining the niche for `Option`s incorrectly where the `Option`'s discriminants don't start at 0.

Rust compiler appears to now prefer a lower niche value, and allows a value to have niches at both the bottom and top of the range.