chore(linter): remove deprecated  `eslint/no-spaced-func` rule (#9482)
This reverts the recent commit that added `eslint/no-spaced-func` as a
new rule.

The reason for removal is that it shouldn't have been added as it is
deprecated by Eslint and we don't want to support deprecated eslint
rules in the linter. See this
[comment](https://github.com/oxc-project/oxc/pull/9392#issuecomment-2690061784).

Rule details here: https://eslint.org/docs/latest/rules/no-spaced-func